requirements.txt
# RetiSense-Multi: Multi-Disease Retinal Screening System
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0

# Image Augmentation
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Metrics & Evaluation
scikit-learn>=1.2.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.22.0

# Mixed Precision Training
accelerate>=0.20.0

# Weights & Biases (optional, for experiment tracking)
wandb>=0.15.0

# Kaggle API (for downloading ODIR-5K)
kaggle>=1.5.0
